Camera: viewports, windowResized -> setViewport.
BREAKING CHANGE: renamed windowResizedAll to setAllViewports,
and windowResized to setViewport.

Added support for x/y components of these functions, so you can
have a camera which uses only part of the screen (minimaps, etc.).

Added Camera.pivot = {xFraction, yFraction} which sets the center
of rotation/scaling for the camera to a fraction of the viewport.

Added `love.graphics.origin` to `applyTransform` so viewport
coordinates are always in pixels.