SceneTree: collectVisible is now a public method.
Since the current `SceneTree.draw` calls `collectVisible` and then
clears the draw order, it would be doing a lot of extra work (and
losing any draw calls that you added during update) if you called it
multiple times per frame. Allowing users to call `collectVisible` lets
them work around this.

Maybe there is a better workaround for this...but it's ok for now.